refactor deletion, handle directories deletion on hover
