+++ catalog/app/containers/Bucket/File/Toolbar/Organize/Context.tsx
@@ -1,18 +1,11 @@
 import invariant from 'invariant'
 import * as React from 'react'
 
-import * as M from '@material-ui/core'
-
-import Code from 'components/Code'
 import * as Bookmarks from 'containers/Bookmarks'
 import type * as Toolbar from 'containers/Bucket/Toolbar'
-import { deleteObject } from 'containers/Bucket/requests'
+import DeleteDialog from 'containers/Bucket/Toolbar/DeleteDialog'
 import * as FileEditor from 'components/FileEditor'
-import * as Notifications from 'containers/Notifications'
-import * as AWS from 'utils/AWS'
 import * as Dialogs from 'utils/Dialogs'
-import Log from 'utils/Logging'
-import * as s3paths from 'utils/s3paths'
 
 export interface OrganizeFileActions {
   toggleBookmark: () => void
@@ -41,56 +34,6 @@ interface OrganizeFileProviderProps {
   editorState: FileEditor.EditorState
   handle: Toolbar.FileHandle
   onReload: () => void
-}
-
-interface DeleteDialogProps {
-  close: () => void
-  handle: Toolbar.FileHandle
-  onReload: () => void
-}
-
-function DeleteDialog({ close, handle, onReload }: DeleteDialogProps) {
-  const [submitting, setSubmitting] = React.useState(false)
-  const s3 = AWS.S3.use()
-  const { push } = Notifications.use()
-
-  const onSubmit = React.useCallback(async () => {
-    setSubmitting(true)
-    try {
-      await deleteObject({ s3, handle })
-      push(`${s3paths.handleToS3Url(handle)} deleted successfully`)
-      close()
-      onReload()
-    } catch (error) {
-      Log.error('Failed to delete file:', error)
-      push(`Failed deleting ${s3paths.handleToS3Url(handle)}`)
-    }
-    setSubmitting(false)
-  }, [s3, handle, push, close, onReload])
-
-  return (
-    <>
-      <M.DialogTitle>Delete object?</M.DialogTitle>
-      <M.DialogContent>
-        <Code>
-          s3://{handle.bucket}/{handle.key}
-        </Code>
-      </M.DialogContent>
-      <M.DialogActions>
-        <M.Button onClick={close} color="primary" variant="outlined">
-          Cancel
-        </M.Button>
-        <M.Button
-          color="primary"
-          disabled={submitting}
-          variant="contained"
-          onClick={onSubmit}
-        >
-          Delete
-        </M.Button>
-      </M.DialogActions>
-    </>
-  )
 }
 
 export function OrganizeFileProvider({
@@ -123,7 +66,7 @@ export function OrganizeFileProvider({
 
   const confirmDelete = React.useCallback(async () => {
     dialogs.open(({ close }) => (
-      <DeleteDialog handle={handle} onReload={onReload} close={close} />
+      <DeleteDialog handles={[handle]} onComplete={onReload} close={close} />
     ))
   }, [dialogs, handle, onReload])
